requirements.txt
# AWS SDK
boto3==1.34.34
botocore==1.34.34
aws-lambda-powertools==2.31.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
pillow==10.2.0

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.17.0
PyMuPDF==1.23.14

# Database - PostgreSQL
sqlalchemy==1.4.49
psycopg2-binary==2.9.9
alembic==1.13.1

# API Framework
fastapi==0.109.0
uvicorn==0.27.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
marshmallow==3.20.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
tenacity==8.2.3
backoff==2.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
moto==4.2.14

# Development
black==24.1.1
flake8==7.0.0
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# ML Libraries
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
img2pdf==0.5.1

# Image Processing
opencv-python==4.9.0.80
reportlab==4.0.9
jinja2==3.1.3

# Redis (for caching)
redis==5.0.1

# Encryption (Fernet for API token storage)
cryptography==42.0.5

# Async HTTP client (Alegra API calls)
httpx==0.27.0

# Logging
structlog==23.2.0

# PostgreSQL Async Support
asyncpg==0.29.0
databases==0.8.0



transformers>=4.30.0
torch>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
databases>=0.8.0 
alembic>=1.13.1
transformers>=4.30.0
torch>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
alembic==1.13.1